Rename schemas/call_signals -> call_signal
