+++ imports/client/components/RTCDebugPage.tsx
@@ -12,7 +12,7 @@ import CallSignals from '../../lib/models/call_signals';
 import Profiles from '../../lib/models/profiles';
 import Puzzles from '../../lib/models/puzzles';
 import { CallParticipantType } from '../../lib/schemas/call_participant';
-import { CallSignalType } from '../../lib/schemas/call_signals';
+import { CallSignalType } from '../../lib/schemas/call_signal';
 import { ProfileType } from '../../lib/schemas/profiles';
 import { PuzzleType } from '../../lib/schemas/puzzles';
