+++ imports/lib/schemas/facade.ts
@@ -1,5 +1,6 @@
 import Announcment from './announcement';
 import CallParticipant from './call_participant';
+import CallSignal from './call_signal';
 import ChatNotifications from './chat_notifications';
 import ChatMessages from './chats';
 import DocumentPermissions from './document_permissions';
@@ -16,6 +17,7 @@ import User from './users';
 const Schemas = {
   Announcment,
   CallParticipant,
+  CallSignal,
   ChatMessages,
   ChatNotifications,
   DocumentPermissions,
